Install ruby gem along with server
This commit manages the mysql ruby gem package whenever the
mysql::server class is declared. This is needed by all the providers
included in this module. The mysql::ruby use the system packages when
avaialable and fall back to using gems if the system package is not able
to be determined.

Spec tests have been provided to test the generated catalog