Make the user in mysql::db to database, not global
Previous to this commit, the mysql::db defined type managed a resource
of type database_user using the value of the user parameter passed in to
mysql::db.  This was problematic since if two different mysql::db
resources were declared with the same user, a duplicate resource error
would occur.

This commit resolves this by assigning the database_user resource to the
database being managed. Note this changes existing behavior and will
likely be cause for a major version release.