(#11565) Don't shell out to set root password
Previously, this commit used an exec that called system mysql commands
with the root password as an argument.  This inadvertently leaks the
password to the process table. This commit creates a script on the
system protected by mode 0700.  The script will be executed by an exec
resource whenever it changes causing the root password to be changed.