Notify mysql restart from set_mysql_rootpw exec
Since the set_mysql_rootpw exec is only created
when the password is being managed, having the mysql
resource (which always exists) reference it leads
to an error when the mysql root password is not being
managed.

This commit resolves the issue by having the set
mysql rootpw exec notify the mysql restart if it
changes.