Allow for the inclusion of arbitrary settings in mysld config
I added a mysqld_misc_settings paramter which allows the inclusion of
arbitrary settings in to the mysqld section of my.cnf.  The
mysqld_misc_settings paramter expects a hash of key value pairs which
correspond to the mysql setting as the key and the value of that
setting as the value.