Fix $etc_root_password by adding conditional to template
The old way would break lots of things because the two templates collided. This
new way should let all configs coexist.

The "before => File[$config_file]" line on setting the root password is
necessary so that the mysqladmin command uses the old configuration before
changing the password; without this, changes to the root password fail on the
first provisioning.