permit use of mysql::db without mysql::server
This removes the hard dependencies on mysql::server from
manifests/db.pp. This permits one to use this module to manage databases
in an existing server, possibly located on a remote host (with a
properly configured /root/.my.cnf).

The dependencies are instead placed in mysql/server.pp and are
implemented using collectors, so that the dependencies are only created
if mysql::server is realized.