Added properties 'ssl_type' and 'ssl_cipher' to mysql_user to enable SSL encrypted authentication.
Added two properties 'ssl_type' and 'ssl_cipher' to mysql_user to allow specifying the required SSL method for the user to log in.
'ssl_type' can take the values '', 'ANY', 'SPECIFIED', or 'X509', which results in no requirement for SSL (''), any SSL-based authentication ('ANY'), SSL-based authentication using a specific cipher ('SPECIFIED'), or authentication using a signed client certificate ('X509').
'ssl_cipher' allows to provide the cipher and requires ssl_type == 'SPECIFIED'.
This is in compliance with the MySQL GRANT syntax, cf. e.g. https://dev.mysql.com/doc/refman/5.6/en/grant.html

The additional properties and corresponding SQL-statements for X509-options are not yet implemented.

Examples:

mysql_user { 'someuser@%':
  ensure        => 'present',
  password_hash => '*F3A2A51A9B0F2BE2468926B4132313728C250DBF',
  ssl_type      => 'ANY'
}

generates

  GRANT USAGE ON *.* TO 'someuser'@'%' REQUIRE SSL;

for an existing user or inserts the REQUIRE-clause in the GRANT command of the create-method for a new user.

mysql_user { 'someuser@%':
  ensure        => 'present',
  password_hash => '*F3A2A51A9B0F2BE2468926B4132313728C250DBF',
  ssl_type      => 'SPECIFIED',
  ssl_cipher    => 'EDH-RSA-DES-CBC3-SHA'
}

generates

  GRANT USAGE ON *.* TO 'someuser'@'%' REQUIRE SSL;
  GRANT USAGE ON *.* TO 'someuser'@'%' REQUIRE CIPHER 'EDH-RSA-DES-CBC3-SHA';

for an existing user or inserts the REQUIRE-clause connected by AND in the GRANT command of the create-method for a new user.