[MODULES-#3557] Add Ubuntu 16.04 package names for language bindings
As in https://tickets.puppetlabs.com/browse/MODULES-3557, Ubuntu 16.04
package names for language bindings are not present, incurring errors
when using the module.

This pull request adds an additional case ('xenial') for the
ruby_package_name parameter in the 'Debian' case, so as to avoid falling
through to the default case. A conditional block for the
php_package_name parameter in the 'Debian' case has also been added, so
as to allow for correct package name usage on 16.04, whilst maintaining
backwards-compatibility with the default case.