MODULES-3904: Abiliity to use SQL_LOG_BIN for MySQL types
There are times where you may want a `mysql_user{}` or mysql_grant{}` to
exist only on one host. The addition of session_sql_log_bin can be set
to enable or disable binary logging for the change. This also works for
`mysql_database{}` but I believe this would be used less often.