(MODULES-4604) conditionalize name validation in mysql_grant type
Users are running into an issue where \`puppet resource mysql_grant\` fails because there is a global name validation using properties that only exist in a compiled catalog and not in the naked type that is instantiated during a \`puppet resource\` run. self[:user] and self[:table] are nil in this use case so there is a failure. This commit surrounds the validation in an if statement so it is not called during \`puppet resource\`. Also, a happy path test is added and the acceptance test is removed."