Fixing empty user/password issue
innobackupex command (which is used by xtrabackup) does not require a username
and a password, so those could be undefined. Since in the mysql::backup::xtrabackup
those were defined as empty strings, and in the template we only check if the
variable is defined, it results in a script that has the --user and --pasword
set, but no actual credentials, which breaks it.

Other backup providers are not affected, since a username and password are required.