(MODULES-5528) mysql_install_db should be optional
Prior to this commit, `mysql_install_db` was a hard requirement in the
mysql provider. Since this command is not in all versions of MySQL, this
commit removes the hard requirement of `mysql_install_db`. Existing
conditional logic already exists to handle situations where the
`mysql_install_db` is not used.