(MODULES-7487) Fix pw comparison MariaDB 10.2.16+
Since MariaDB 10.2.16, the `SET PASSWORD` command stores user passwords
in the `mysql_native_password` plugin `authentication_string` instead of
in the `password` column:

```
+------+-----------+----------+-----------------------+-------------------------------------------+
| user | host      | password | plugin                | authentication_string                     |
+------+-----------+----------+-----------------------+-------------------------------------------+
| root | localhost |          | mysql_native_password | *01396341988BCA5088C3A5DB5D7E434947096D4F |
+------+-----------+----------+-----------------------+-------------------------------------------+
```

The `puppetlabs/mysql` checks the `password` column instead of the
`authentication_string` on all versions of MariaDB, resulting in every
Puppet run attempting to update managed user passwords even when they
haven't changed.

Update the if statement to check the plugin authentication string on
MariaDB 10.2.16 and newer.