Update README.md Ubuntu example to xenial
Hi - thanks for this great forge module. It got me over the line when I was forced to use mariadb

Just reviewing my code, and the service name on xenial is mysqld, not mysql, so I thought I'd take the opportunity to refresh this part of the doc, as I copy/pasted verbatim!