Putting Debian back into the metadata.json
Currently we have raised an issue to track Debian failures https://tickets.puppetlabs.com/browse/IAC-1595
In order to stop blocking us we planned on stopping to test on these OSs temporarily and the obvious move was to remove from the metadata.json

However the problem with this is, when we cut a release the Forge will use this to show the supported OSs and we do support Debian, just we wanted to disable testing. We will have to come up with a 'better' was to disable tests on specific OSs.