(MAINT) Fix failing Ubuntu test cases
Prior to this commit, the Ubuntu 20.04 spec testing was failing two
test cases at mysql_grant_spec. One of them was a simple mismatch
between the expected stdout and actual stdout of the test case, and the
other had to do with a misbehaviour from the module at the time of
granting privileges to users. The module, upon instructed to grant 'ALL'
pivileges to the user, creates and array that contains both 'ALL' and
all dynamic privileges and passes that array as the parameter for
privilege granting, causing an error in idempotency.

This commit aims to the previously mentioned errors. The first error is
a simple 'expected message' adjust whereas for the second issue, a
temporary workaround that removes the duplicated dynamic privileges
reference has been implemented.

In the future, the issue should be assessed and a fix that actually
modifies the behaviour of the module to avoid such duplication should be
implemented.