(Bugfix) Grant privileges idempotency Fix
Prior to this commit, the Ubuntu 20.04 spec testing was failing two test cases at mysql_grant_spec. One of them was a simple mismatch between the expected stdout and actual stdout of the test case, and the other had to do with a misbehavior from the module at the time of granting privileges to users.
The module upon retrieving the current privileges accounted only for the original static privileges and not the dynamic privileges added in newer mysql versions, causing an error in the code that is used to condense the array of privileges that make up `ALL` into a single statement.
The code which condenses the returned privileges into a simple `ALL`  has now been updated correctly trigger when dynamic privileges are in effect, avoiding this issue and ensuring idempotency.