(MAINT) Fix duplicate procedure/function tests
Prior to this commit there were two test cases at mysql_grant_spec that
would throw an error when run locally. The issue with these tests was
that, upon running the code, they would fail because the function or
procedure they were trying to create already existed.

This commit adds a 'DROP ... IF EXISTS name' SQL command right before
the test command is executed in order to clean up existing functions /
procedures that may interfere with the correct behaviour of the test.