(IAC-1595) MySQL maintenance
Prior to this commit, MySQL module had the entirety of the Debian OS
family disabled due to provisioning issues at cloud level. JIRA ticket
IAC-1595 reported this behaviour. In addition, Debian 11 was reported
to have an unfinished implementation.

This commit was aimed at re-enabling the debian systems and
troubleshooting the failing ones. As well as adding the missing
configuration for Debian 11.

During this commit, it was established that the best way to troubleshoot
the debian provisioning issues was to switch the provisioning service to
GCP. Upon fixing the provisioning issues, a number of test error cases
related to privileges and bad locale variables were reported and those
were addressed as well.

Finally, some minor maintenance changes were applied as well.