Harden service class
Prior to this commit the variable `mysqlsocket` was passed to the `exec` resource
in such a way that could allow unsafe executions on the remote host.

This commit fixes the above by properly parameterizing the arguments passed
to the `command` and `unless` parameters of the `exec` resource.

The variable is also wrapped with a `shell_escape` for good measure.