RHEL9 Fixes
Prior to this commit, xtrabackup acceptance tests would fail when the
OS was RHEL 9. This is because xtrabackup is not currently supported for
RHEL 9.

This commit moves the xtrabackup acceptance tests in to their own
spec file and adds conditions to the tests that ensure they are only
executed on supported OSes.