Disable mysqld apparmor profiles
Prior to this commit on certian OSes the mysqld apparmor profile present
on the github runner would prevent the mysql service from properly
starting inside our litmusimage container.

The follow OSes were impacted for both puppet 6 & 7 runs:

* debian 9
* debian 10
* ubuntu 20.04

This commit adds a new step to the `pr_test` and `nightly` workflows
that disables the mysqld apparmor profile if the current provisioner is
`provision::docker`.

While this fixes the issue, it does make me wonder if this is the right
approach.