Use consistent values for ssl-ca, ssl-cert and ssl-key
The default value of the server's ssl-ca, ssl-cert and ssl-key
parameters are system-dependant:
  * Some systems set these parameters to `undef`;
  * Some systems set these parameters to a filename which is not managed
    by the module.

Recent versions of MariaDB insist on being able to reatd the files these
parameters point to even if TLS is not in use.  As a consequence,
MariaDB is broken and cannot start successfuly.

Always use `undef` as the default version of these parameters so that
users have the same experience regardless of the Operating System they
are runnning.